Implemented ConfigPresenter.

+++ dotnet/Tricycle.UI/Models/AudioQualityPreset.cs
@@ -5,7 +5,7 @@ namespace Tricycle.UI.Models
         public string Key { get; set; }
         public string Format { get; set; }
         public string Mixdown { get; set; }
-        public int? Quality { get; set; }
+        public decimal? Quality { get; set; }
         public bool RemoveDisabled { get; set; }
     }
 }
